big changes

+++ src/hr/fer/zemris/structures/Pair.java
@@ -3,8 +3,8 @@ package hr.fer.zemris.structures;
 
 public class Pair implements Comparable<Pair>{
 	public double value;
-	public long id;
-	public Pair(long id, double value)
+	public int id;
+	public Pair(int id, double value)
 	{
 		this.id = id;
 		this.value = value;
@@ -18,7 +18,7 @@ public class Pair implements Comparable<Pair>{
 	@Override
 	public int compareTo(Pair o) 
 	{
-		return Double.compare(value, o.value)==0?Long.compare(id, o.id):Double.compare(value, o.value);
+		return Double.compare(value, o.value)==0?Integer.compare(id, o.id):Double.compare(value, o.value);
 	}
 	
 }
