+++ src/hr/fer/zemris/structures/dot/DotCache.java
@@ -7,16 +7,16 @@ public class DotCache implements Serializable{
 	 * 
 	 */
 	private static final long serialVersionUID = 8330283207842965329L;
-	private long id;
+	private int id;
 	private int component;
 	private double val;
-	public DotCache(long id, int component, double val)
+	public DotCache(int id, int component, double val)
 	{
 		this.id = id;
 		this.component = component;
 		this.val = val;
 	}
-	public long getId()
+	public int getId()
 	{
 		return this.id;
 	}
